refactor: separate out the logic for filtering component list by version
